+++ pkg/cluster/internal/validate/aws.go
@@ -46,6 +46,7 @@ var AWSNodeImageFormat = "ami-[IMAGE_ID]"
 func validateAWS(spec commons.KeosSpec, providerSecrets map[string]string) error {
 	var err error
 	var ctx = context.TODO()
+	deviceRegex := regexp.MustCompile(commons.DeviceNameRegex)
 
 	cfg, err := commons.AWSGetConfig(ctx, providerSecrets, spec.Region)
 	if err != nil {
@@ -104,6 +105,42 @@ func validateAWS(spec commons.KeosSpec, providerSecrets map[string]string) error
 		if err := validateVolumeType(spec.ControlPlane.RootVolume.Type, AWSVolumes); err != nil {
 			return errors.Wrap(err, "spec.control_plane.root_volume: Invalid value: \"type\"")
 		}
+		if spec.ControlPlane.CRIVolume.Enabled != nil {
+			if !*spec.ControlPlane.CRIVolume.Enabled {
+				defaultCRIVolume := commons.CustomVolume{}
+				defaultCRIVolume.Enabled = commons.ToPtr(false)
+				if !reflect.DeepEqual(spec.ControlPlane.CRIVolume, defaultCRIVolume) {
+					return errors.New("spec.control_plane.cri_volume: Invalid state: no other fields should be specified if enabled is false")
+				}
+			} else {
+				if err := validateVolumeType(spec.ControlPlane.CRIVolume.Type, AWSVolumes); err != nil {
+					return errors.Wrap(err, "spec.control_plane.cri_volume: Invalid value: \"type\"")
+				}
+
+				if !reflect.DeepEqual(spec.ControlPlane.CRIVolume, commons.CustomVolume{}) && spec.ControlPlane.CRIVolume.Size < 8 {
+					return errors.New("spec.control_plane.cri_volume: Invalid value: \"size\": must be greater or equal to 8")
+				}
+			}
+		}
+
+		if spec.ControlPlane.ETCDVolume.Enabled != nil {
+			if !*spec.ControlPlane.ETCDVolume.Enabled {
+				defaultETCDVolume := commons.CustomVolume{}
+				defaultETCDVolume.Enabled = commons.ToPtr(false)
+				if !reflect.DeepEqual(spec.ControlPlane.ETCDVolume, defaultETCDVolume) {
+					return errors.New("spec.control_plane.etcd_volume: Invalid state: no other fields should be specified if enabled is false")
+				}
+			} else {
+
+				if err := validateVolumeType(spec.ControlPlane.ETCDVolume.Type, AWSVolumes); err != nil {
+					return errors.Wrap(err, "spec.control_plane.etcd_volume: Invalid value: \"type\"")
+				}
+
+				if !reflect.DeepEqual(spec.ControlPlane.ETCDVolume, commons.CustomVolume{}) && spec.ControlPlane.ETCDVolume.Size < 8 {
+					return errors.New("spec.control_plane.etcd_volume: Invalid value: \"size\": must be greater or equal to 8")
+				}
+			}
+		}
 
 		for i, ev := range spec.ControlPlane.ExtraVolumes {
 			if ev.DeviceName == "" {
@@ -112,6 +149,13 @@ func validateAWS(spec commons.KeosSpec, providerSecrets map[string]string) error
 			if err := validateVolumeType(ev.Type, AWSVolumes); err != nil {
 				return errors.Wrap(err, "spec.control_plane.extra_volumes["+strconv.Itoa(i)+"]: Invalid value: \"type\"")
 			}
+			if ev.Size < 8 {
+				return errors.New("spec.control_plane.extra_volumes[" + strconv.Itoa(i) + "]: Invalid value: \"size\": must be greater or equal to 8")
+			}
+			if !deviceRegex.MatchString(ev.DeviceName) {
+				return errors.New("spec.control_plane.extra_volumes[" + strconv.Itoa(i) + "]: Invalid value: \"device_name\": invalid name, must be ^/dev/(sd[a-z]|xvd([a-d][a-z]|[e-z]))$ format")
+			}
+
 			for j, ev2 := range spec.ControlPlane.ExtraVolumes {
 				if i != j {
 					if ev.DeviceName == ev2.DeviceName {
@@ -143,12 +187,35 @@ func validateAWS(spec commons.KeosSpec, providerSecrets map[string]string) error
 		if err := validateVolumeType(wn.RootVolume.Type, AWSVolumes); err != nil {
 			return errors.Wrap(err, "spec.worker_nodes."+wn.Name+".root_volume: Invalid value: \"type\"")
 		}
+		if wn.CRIVolume.Enabled != nil {
+			if !*wn.CRIVolume.Enabled {
+				defaultCRIVolume := commons.CustomVolume{}
+				defaultCRIVolume.Enabled = commons.ToPtr(false)
+				if !reflect.DeepEqual(wn.CRIVolume, defaultCRIVolume) {
+					return errors.New("spec.worker_nodes." + wn.Name + ".cri_volume: Invalid state: no other fields should be specified if enabled is false")
+				}
+			} else {
+				if err := validateVolumeType(wn.CRIVolume.Type, AWSVolumes); err != nil {
+					return errors.Wrap(err, "spec.worker_nodes."+wn.Name+".cri_volume: Invalid value: \"type\"")
+				}
+				if wn.CRIVolume.Size < 8 {
+					return errors.New("spec.worker_nodes." + wn.Name + ".cri_volume: Invalid value: \"size\": must be greater or equal to 8")
+				}
+			}
+		}
+
 		for i, ev := range wn.ExtraVolumes {
 			if ev.DeviceName == "" {
 				return errors.New("spec.worker_nodes." + wn.Name + ".extra_volumes[" + strconv.Itoa(i) + "]: Invalid value: \"device_name\": is required")
 			}
 			if err := validateVolumeType(ev.Type, AWSVolumes); err != nil {
 				return errors.Wrap(err, "spec.worker_nodes."+wn.Name+".extra_volumes["+strconv.Itoa(i)+"]: Invalid value: \"type\"")
+			}
+			if ev.Size < 8 {
+				return errors.New("spec.worker_nodes." + wn.Name + ".extra_volumes[" + strconv.Itoa(i) + "]: Invalid value: \"size\": must be greater or equal to 8")
+			}
+			if !deviceRegex.MatchString(ev.DeviceName) {
+				return errors.New("spec.worker_nodes." + wn.Name + ".extra_volumes[" + strconv.Itoa(i) + "]: Invalid value: \"device_name\": invalid name, must be ^/dev/(sd[a-z]|xvd([a-d][a-z]|[e-z]))$ format")
 			}
 			for j, ev2 := range spec.ControlPlane.ExtraVolumes {
 				if i != j {