+++ pkg/commons/cluster.go
@@ -29,6 +29,16 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
+var EtcdVolumeSize = 8
+var CriVolumeSize = 128
+var RootVolumeDefaultSize = 128
+var RootVolumeManagedDefaultSize = 256
+var DeviceNameRegex = "^/dev/(sd[a-z]|xvd([a-d]|[a-d][a-z]|[e-z]))$"
+
+var AWSVolumeType = "gp3"
+var AzureVMsVolumeType = "Standard_LRS"
+var GCPVMsVolumeType = "pd-standard"
+
 type Resource struct {
 	APIVersion string      `yaml:"apiVersion" validate:"required"`
 	Kind       string      `yaml:"kind" validate:"required"`
@@ -110,21 +120,25 @@ type KeosSpec struct {
 
 	Keos Keos `yaml:"keos,omitempty"`
 
-	ControlPlane struct {
-		Managed         bool                `yaml:"managed" validate:"boolean"`
-		NodeImage       string              `yaml:"node_image,omitempty"`
-		HighlyAvailable *bool               `yaml:"highly_available,omitempty" validate:"boolean"`
-		Size            string              `yaml:"size,omitempty" validate:"required_if=Managed false"`
-		RootVolume      RootVolume          `yaml:"root_volume,omitempty"`
-		Tags            []map[string]string `yaml:"tags,omitempty"`
-		AWS             AWSCP               `yaml:"aws,omitempty"`
-		Azure           AzureCP             `yaml:"azure,omitempty"`
-		ExtraVolumes    []ExtraVolume       `yaml:"extra_volumes,omitempty" validate:"dive"`
-	} `yaml:"control_plane"`
+	ControlPlane Controlplane `yaml:"control_plane" validate:"required,dive"`
 
 	WorkerNodes WorkerNodes `yaml:"worker_nodes" validate:"required,dive"`
 
 	ClusterConfigRef ClusterConfigRef `yaml:"cluster_config_ref,omitempty" validate:"dive"`
+}
+
+type Controlplane struct {
+	Managed         bool                `yaml:"managed" validate:"boolean"`
+	NodeImage       string              `yaml:"node_image,omitempty"`
+	HighlyAvailable *bool               `yaml:"highly_available,omitempty" validate:"boolean"`
+	Size            string              `yaml:"size,omitempty" validate:"required_if=Managed false"`
+	RootVolume      RootVolume          `yaml:"root_volume,omitempty"`
+	Tags            []map[string]string `yaml:"tags,omitempty"`
+	AWS             AWSCP               `yaml:"aws,omitempty"`
+	Azure           AzureCP             `yaml:"azure,omitempty"`
+	CRIVolume       CustomVolume        `yaml:"cri_volume,omitempty"  validate:"dive"`
+	ETCDVolume      CustomVolume        `yaml:"etcd_volume,omitempty"  validate:"dive"`
+	ExtraVolumes    []ExtraVolume       `yaml:"extra_volumes,omitempty" validate:"dive"`
 }
 
 type Keos struct {
@@ -184,6 +198,7 @@ type WorkerNodes []struct {
 	NodeGroupMaxSize int               `yaml:"max_size,omitempty" validate:"omitempty,required_with=NodeGroupMinSize,numeric"`
 	NodeGroupMinSize *int              `yaml:"min_size,omitempty" validate:"omitempty,required_with=NodeGroupMaxSize,numeric,gte=0"`
 	RootVolume       RootVolume        `yaml:"root_volume,omitempty"`
+	CRIVolume        CustomVolume      `yaml:"cri_volume,omitempty"  validate:"dive"`
 	ExtraVolumes     []ExtraVolume     `yaml:"extra_volumes,omitempty" validate:"dive"`
 }
 
@@ -205,12 +220,36 @@ type RootVolume struct {
 type ExtraVolume struct {
 	Name          string `yaml:"name,omitempty"`
 	DeviceName    string `yaml:"device_name,omitempty"`
+	Size          int    `yaml:"size,omitempty"`
+	Type          string `yaml:"type,omitempty"`
+	Label         string `yaml:"label,omitempty"`
+	Encrypted     bool   `yaml:"encrypted,omitempty" validate:"boolean"`
+	EncryptionKey string `yaml:"encryption_key,omitempty"`
+	MountPath     string `yaml:"mount_path,omitempty"`
+}
+
+type CustomVolume struct {
+	Enabled       *bool  `yaml:"enabled,omitempty"`
+	Size          int    `yaml:"size,omitempty"`
+	Type          string `yaml:"type,omitempty"`
+	Encrypted     bool   `yaml:"encrypted,omitempty"`
+	EncryptionKey string `yaml:"encryption_key,omitempty"`
+}
+
+type ETCDVolume struct {
 	Size          int    `yaml:"size" validate:"required,numeric"`
 	Type          string `yaml:"type,omitempty"`
 	Label         string `yaml:"label" validate:"required"`
 	Encrypted     bool   `yaml:"encrypted,omitempty" validate:"boolean"`
 	EncryptionKey string `yaml:"encryption_key,omitempty"`
-	MountPath     string `yaml:"mount_path" validate:"required"`
+}
+
+type CRIVolume struct {
+	Size          int    `yaml:"size" validate:"required,numeric"`
+	Type          string `yaml:"type,omitempty"`
+	Label         string `yaml:"label" validate:"required"`
+	Encrypted     bool   `yaml:"encrypted,omitempty" validate:"boolean"`
+	EncryptionKey string `yaml:"encryption_key,omitempty"`
 }
 
 type ClusterCredentials struct {
@@ -391,6 +430,106 @@ func (s KeosSpec) Init() KeosSpec {
 	return s
 }
 
+func (s KeosSpec) InitVolumes() KeosSpec {
+	var volumeType string
+
+	switch s.InfraProvider {
+	case "aws":
+		volumeType = AWSVolumeType
+
+		if !s.ControlPlane.Managed {
+
+			if s.ControlPlane.CRIVolume.Enabled == nil || *s.ControlPlane.CRIVolume.Enabled {
+				s.ControlPlane.CRIVolume.Enabled = ToPtr(true)
+				s = initControlPlaneCRIVolume(s, volumeType)
+			}
+			if s.ControlPlane.ETCDVolume.Enabled == nil || *s.ControlPlane.ETCDVolume.Enabled {
+				s.ControlPlane.ETCDVolume.Enabled = ToPtr(true)
+				s = initControlPlaneETCDVolume(s, volumeType)
+			}
+			s = initControlPlaneRootVolume(s, volumeType, !*s.ControlPlane.CRIVolume.Enabled)
+		}
+
+		for i := range s.WorkerNodes {
+
+			if s.WorkerNodes[i].CRIVolume.Enabled == nil || *s.WorkerNodes[i].CRIVolume.Enabled {
+				s.WorkerNodes[i].CRIVolume.Enabled = ToPtr(true)
+				checkAndFill(&s.WorkerNodes[i].CRIVolume.Size, CriVolumeSize)
+				checkAndFill(&s.WorkerNodes[i].CRIVolume.Type, volumeType)
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Size, RootVolumeDefaultSize)
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Type, volumeType)
+			} else {
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Size, RootVolumeManagedDefaultSize)
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Type, volumeType)
+			}
+
+		}
+
+	case "gcp":
+		if !s.ControlPlane.Managed {
+			volumeType = GCPVMsVolumeType
+			if s.ControlPlane.CRIVolume.Enabled == nil || *s.ControlPlane.CRIVolume.Enabled {
+				s.ControlPlane.CRIVolume.Enabled = ToPtr(true)
+				s = initControlPlaneCRIVolume(s, volumeType)
+			}
+			if s.ControlPlane.ETCDVolume.Enabled == nil || *s.ControlPlane.ETCDVolume.Enabled {
+				s.ControlPlane.ETCDVolume.Enabled = ToPtr(true)
+				s = initControlPlaneETCDVolume(s, volumeType)
+			}
+			s = initControlPlaneRootVolume(s, volumeType, !*s.ControlPlane.CRIVolume.Enabled)
+		}
+		for i := range s.WorkerNodes {
+			if !s.ControlPlane.Managed {
+				if s.WorkerNodes[i].CRIVolume.Enabled == nil || *s.WorkerNodes[i].CRIVolume.Enabled {
+					s.WorkerNodes[i].CRIVolume.Enabled = ToPtr(true)
+					checkAndFill(&s.WorkerNodes[i].CRIVolume.Size, CriVolumeSize)
+					checkAndFill(&s.WorkerNodes[i].CRIVolume.Type, volumeType)
+					checkAndFill(&s.WorkerNodes[i].RootVolume.Size, RootVolumeDefaultSize)
+					checkAndFill(&s.WorkerNodes[i].RootVolume.Type, volumeType)
+				}
+			} else {
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Size, RootVolumeManagedDefaultSize)
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Type, volumeType)
+			}
+
+		}
+
+	case "azure":
+		if !s.ControlPlane.Managed {
+			volumeType = AzureVMsVolumeType
+			if s.ControlPlane.CRIVolume.Enabled == nil || *s.ControlPlane.CRIVolume.Enabled {
+				s.ControlPlane.CRIVolume.Enabled = ToPtr(true)
+				s = initControlPlaneCRIVolume(s, volumeType)
+			}
+			if s.ControlPlane.ETCDVolume.Enabled == nil || *s.ControlPlane.ETCDVolume.Enabled {
+				s.ControlPlane.ETCDVolume.Enabled = ToPtr(true)
+				s = initControlPlaneETCDVolume(s, volumeType)
+			}
+			s = initControlPlaneRootVolume(s, volumeType, !*s.ControlPlane.CRIVolume.Enabled)
+		}
+
+		for i := range s.WorkerNodes {
+
+			if !s.ControlPlane.Managed {
+				if s.WorkerNodes[i].CRIVolume.Enabled == nil || *s.WorkerNodes[i].CRIVolume.Enabled {
+					s.WorkerNodes[i].CRIVolume.Enabled = ToPtr(true)
+					checkAndFill(&s.WorkerNodes[i].CRIVolume.Size, CriVolumeSize)
+					checkAndFill(&s.WorkerNodes[i].CRIVolume.Type, volumeType)
+					checkAndFill(&s.WorkerNodes[i].RootVolume.Size, RootVolumeDefaultSize)
+					checkAndFill(&s.WorkerNodes[i].RootVolume.Type, volumeType)
+				}
+			} else {
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Size, RootVolumeManagedDefaultSize)
+				checkAndFill(&s.WorkerNodes[i].RootVolume.Type, volumeType)
+			}
+		}
+
+	}
+
+	return s
+
+}
+
 // Read descriptor file
 func GetClusterDescriptor(descriptorPath string) (*KeosCluster, *ClusterConfig, error) {
 	var keosCluster KeosCluster
@@ -432,7 +571,7 @@ func GetClusterDescriptor(descriptorPath string) (*KeosCluster, *ClusterConfig, 
 				if err != nil {
 					return nil, nil, err
 				}
-
+				keosCluster.Spec = keosCluster.Spec.InitVolumes()
 				err = validate.Struct(keosCluster)
 				if err != nil {
 					return nil, nil, err